include upload/update/delete for images with uploads

+++ app/__init__.py
@@ -33,12 +33,20 @@ app.cli.add_command(seed_commands)
 app.config.from_object(Config)
 
 UPLOAD_FOLDER = 'uploads'
+IMAGE_FOLDER = 'images'
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
-app.config['ALLOWED_EXTENSIONS'] = {'mp3', 'wav', 'flac'}  
+app.config['IMAGE_FOLDER'] = IMAGE_FOLDER
+app.config['ALLOWED_EXTENSIONS'] = {'mp3', 'wav', 'flac'}
+app.config['ALLOWED_IMAGES']  = {'png', 'jpeg'}
+
+## UPLOADS ##
 
 #Boolean Checks If File Is Allowed
 def allowed_file(filename):
     return '.' in filename and filename.rsplit('.', 1)[1].lower() in app.config['ALLOWED_EXTENSIONS']
+
+def allowed_image(image):
+    return '.' in image and image.rsplit('.', 1)[1].lower() in app.config['ALLOWED_IMAGES']
 
 #Saves Approved Files To Upload Folder through app.config
 def save_song(file, filename):
@@ -54,7 +62,22 @@ def save_song(file, filename):
             print(f"FILE PATH: {file_path} doesn't exist.") 
     else:
         return None
-    
+
+def save_image(image, image_name):
+    if image_name and allowed_image(image_name):
+        secured_filename = secure_filename(image_name) 
+        upload_file_path = app.config['IMAGE_FOLDER']
+        file_path = os.path.join(upload_file_path, secured_filename)
+        os.makedirs(upload_file_path, exist_ok=True)
+        image.save(file_path)        
+        if os.path.exists(file_path):
+            print(f"FILE PATH: {file_path} does exist !!")
+            return file_path
+        else:
+            print(f"FILE PATH: {file_path} doesn't exist.") 
+    else:
+        return None
+
 #Upload Route
 @app.route('/api/songs/upload', methods=['POST', 'GET'])
 def upload_song():
@@ -63,9 +86,13 @@ def upload_song():
         return render_template('upload_song.html', form=form)
     
     form = UploadSongForm()
-    file = request.files['filename']   
+    file = request.files['filename'] 
     filename = file.filename            
     file_path = save_song(file, filename)
+
+    image = request.files['image']
+    image_name = image.filename
+    image_path = save_image(image, image_name)
 
     if file_path:
         song = Song(
@@ -74,8 +101,9 @@ def upload_song():
             artist = form.data['artist'],
             album = form.data['album'],
             genre = form.data['genre'],
-            image = form.data['image'],
-            file_path = file_path
+            image = image_name,
+            file_path = file_path,
+            image_path = image_path
         )        
         db.session.add(song)
         db.session.commit()
@@ -94,19 +122,23 @@ def update_song_information(id):
         return render_template('update_song.html', form=form)
     
     if request.method in ['PUT', 'POST'] and form.validate_on_submit():
-        
         file = request.files['filename']   
         filename = file.filename         
         file_path = save_song(file, filename)
-        
+
+        image = request.files['image']
+        image_name = image.filename
+        image_path = save_image(image, image_name)
+
         song_to_update = Song.query.get(id)
         song_to_update.filename = filename
         song_to_update.title = form.data['title']
         song_to_update.artist = form.data['artist']
         song_to_update.album = form.data['album']
         song_to_update.genre = form.data['genre']
-        song_to_update.image = form.data['image']
+        song_to_update.image = image_name
         song_to_update.file_path = file_path
+        song_to_update.image_path = image_path
         db.session.commit()
         print(f"Updated Song in database.")
         return song_to_update.to_dict()
@@ -133,24 +165,29 @@ def audio_player():
     file_names = os.listdir(file_path)
     return render_template('audio_player.html', file_names=file_names)
 
-#Delete song from /uploads and DB 
-@app.route('/api/songs/<int:id>/delete', methods=['DELETE'])
+#Delete song from /uploads and DB w/ associated image
+@app.route('/api/songs/<int:id>/delete', methods=['GET','DELETE'])
 def delete_from_uploads_and_db(id):
-    song_for_db = Song.query.get(id)
-    song = Song.query.get(id).filename
-    transformed_song = song.replace(' ', '_').replace(',', '').replace("'", '')
-    
-    root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
-    file_path = root + '/' + app.config['UPLOAD_FOLDER'] + '/' + transformed_song
-    os.remove(file_path)
+    if request.method == "GET":
+        song_for_db = Song.query.get(id)
+        song = Song.query.get(id).filename
+        transformed_song = song.replace(' ', '_').replace(',', '').replace("'", '')
+        
+        root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
+        file_path = root + '/' + app.config['UPLOAD_FOLDER'] + '/' + transformed_song
+        os.remove(file_path)
 
-    db.session.delete(song_for_db)
-    db.session.commit()
-    print(f"{song} was succesfully deleted from uploads and from database.")
-    return f"{song} was successfully deleted."
-    
-    
-    
+        image = Song.query.get(id).image
+        transformed_image = image.replace(' ', '_').replace(',', '').replace("'", '')
+
+        image_path = root + '/' + app.config['IMAGE_FOLDER'] + '/' + transformed_image
+        os.remove(image_path)
+
+        db.session.delete(song_for_db)
+        db.session.commit()
+        print(f"{song} was succesfully deleted from uploads and from database with image.")
+        return f"{song} was successfully deleted."
+        
 #Looping through uploads 
 @app.route('/api/config')
 def see_config():
@@ -160,6 +197,14 @@ def see_config():
         print(file)
     
     return f"Contents of {upload_folder_path}: {files}"
+
+app.route('/api/files')
+def see_request_files():
+    for key, file_storage in request.files.items():
+        print(f"File field name: {key}")
+        print(f"File name: {file_storage.filename}")
+        print(f"Content type: {file_storage.content_type}")
+
 
 app.register_blueprint(user_routes, url_prefix='/api/users')
 app.register_blueprint(auth_routes, url_prefix='/api/auth')
